refactor: tidy virtual page and add lang switch
